Move fixed size conformance to separate files
